Refactor to conform more closely to MVC
Changed the routing so requests weren't being sent to the same endpoint.
Also changed the file structure and location of various elements to get
closer to MVC architecture. Still need to do a bit of reading to
understand the concept more fully and implement.
Changed the delete from input/submit to button/submit, this allowed
me to make the request body more conventionally constructed,
which allowed for a slightly simpler delete function.